+++ tsup.config.ts
@@ -3,7 +3,7 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.ts"],
+  entry: ["src/index.ts", "src/preset.ts"],
   format: ["esm", "cjs"],
   dts: true,
   sourcemap: true,